IFEDMethodBase, part 8.
